+++ Camera/FrontCamera.cpp
@@ -68,7 +68,7 @@ void FrontCamera::setCamera()
     glLightfv(GL_LIGHT0, GL_POSITION, m_lightPosition);
 }
 
-void FrontCamera::setReferenceImage(const char *path,GLuint imageID,Vector &position,size_t width,size_t height)
+void FrontCamera::setReferenceImage(const char *path,GLuint imageID,Vector &position,unsigned int width,unsigned int height)
 {
     if(frontReference)
     {
@@ -80,7 +80,7 @@ void FrontCamera::setReferenceImage(const char *path,GLuint imageID,Vector &posi
     frontReference->isShow=true;
 }
 
-void FrontCamera::getViewportImage(GLuint &texID,std::string &path,Vector &position,size_t &width,size_t &height)
+void FrontCamera::getViewportImage(GLuint &texID,std::string &path,Vector &position,unsigned int &width,unsigned int &height)
 {
     if(frontReference)
     {
@@ -105,15 +105,15 @@ void FrontCamera::disableReference()
     frontReference->isShow=false;
 }
 
-void FrontCamera::setCameraForSelectionD(size_t x1,size_t y1,size_t x2,size_t y2,size_t h)
+void FrontCamera::setCameraForSelectionD(unsigned int x1,unsigned int y1,unsigned int x2,unsigned int y2,unsigned int h)
 {
     glViewport((GLint)m_startX,(GLint)m_startY,(GLint)m_width,(GLint)m_height);
     GLint viewport[4];
     glGetIntegerv (GL_VIEWPORT, viewport);
     glMatrixMode(GL_PROJECTION);
     glLoadIdentity();
-    size_t sw=(x2-x1)>5?(x2-x1):5;
-    size_t sh=(y2-y1)>5?(y2-y1):5;
+    unsigned int sw=(x2-x1)>5?(x2-x1):5;
+    unsigned int sh=(y2-y1)>5?(y2-y1):5;
     gluPickMatrix((GLdouble) (x1+x2)/2,(GLdouble) (h - (y1+y2)/2), (GLdouble)sw,(GLdouble)sh, viewport);
     glOrtho(-m_width*(m_eye.y)*0.005f,m_width*(m_eye.y)*0.005f,-m_height*(m_eye.y)*0.005f,m_height*(m_eye.y)*0.005f,m_nearPlane,m_farPlane);
     glMatrixMode(GL_MODELVIEW);
