+++ History/Log_EdgeEndChange.cpp
@@ -1,6 +1,6 @@
 #include "Log_EdgeEndChange.h"
 
-Log_EdgeEndChange::Log_EdgeEndChange(size_t theTarget,size_t theIndex,int theEnd):HistoryLog(LogType::Edge_End_Change,theTarget),index(theIndex),end(theEnd)
+Log_EdgeEndChange::Log_EdgeEndChange(unsigned int theTarget,unsigned int theIndex,int theEnd):HistoryLog(LogType::Edge_End_Change,theTarget),index(theIndex),end(theEnd)
 {
 }
