+++ History/Log_EdgeRightChange.cpp
@@ -1,6 +1,6 @@
 #include "Log_EdgeRightChange.h"
 
-Log_EdgeRightChange::Log_EdgeRightChange(size_t theTarget,size_t theIndex,size_t theRight):HistoryLog(LogType::Edge_Right_Change,theTarget),index(theIndex),right(theRight)
+Log_EdgeRightChange::Log_EdgeRightChange(unsigned int theTarget,unsigned int theIndex,unsigned int theRight):HistoryLog(LogType::Edge_Right_Change,theTarget),index(theIndex),right(theRight)
 {
 }
